Table of PICKMDL criteria

Function crit_table takes several x13 output objects as input and produces a table of criteria (matrix class). The other functions are underlying functions that take a single x13 output object as input.

Usage

crit_table(sa_list)

crit123_m_aic(sa)

crit1(sa)

crit2(sa)

crit3(sa)

m_aic(sa)

Arguments

sa_list

List of several x13 output objects. That is, spec can be output from x13_multi.

sa

A single x13 output object.

Value

A matrix, a vector or a single numerical value

Examples

myseries <- pickmdl_data("myseries")

spec5 <- x13_spec_pickmdl(spec = "RSA3", transform.function = "Log")

sa5 <- x13_multi(myseries, spec = spec5)   

crit_table(sa5)
#>           crit1      crit2      crit3    m_aic
#> [1,] 0.04432100 0.03914802 -0.8365831 1248.975
#> [2,] 0.03697842 0.11701685 -0.7158467 1207.235
#> [3,] 0.04003758 0.65585839  0.0000000 1192.545
#> [4,] 0.04316176 0.27888160 -0.9697126 1247.015
#> [5,] 0.03711805 0.48234261 -0.3487343 1186.795

crit123_m_aic(sa5[[4]])
#>         crit1         crit2         crit3         m_aic 
#>    0.04316176    0.27888160   -0.96971259 1247.01527314 

crit1(sa5[[4]])
#> [1] 0.04316176
crit2(sa5[[4]])
#> [1] 0.2788816
crit3(sa5[[4]])
#> [1] -0.9697126
m_aic(sa5[[4]]) 
#> [1] 1247.015