x13 with PICKMDL and partial concurrent possibilities

x13 can be run as usual (automdl) or with a PICKMDL specification. The ARIMA model, outliers and filters can be identified at a certain date and then held fixed (with a new outlier-span).

Usage

x13_pickmdl(
  series,
  spec,
  corona = FALSE,
  ...,
  pickmdl_method = "first",
  star = 1,
  when_star = warning,
  when_automdl = message,
  when_finalnotok = NULL,
  identification_end = NULL,
  identification_estimate.to = NULL,
  identify_t_filter = FALSE,
  identify_s_filter = FALSE,
  identify_outliers = TRUE,
  identify_arima_mu = TRUE,
  automdl.enabled = FALSE,
  fastfirst = TRUE,
  verbose = FALSE,
  output = "sa",
  add_comment = TRUE
)

x13_automdl(..., automdl.enabled = TRUE)

Arguments

series

x13 parameter

spec

An x13_spec output object or a list of several objects as outputted from x13_spec_pickmdl. In the case of a single object and when automdl.enabled is FALSE, spec will be converted internally by x13_spec_pickmdl with default five arima model specifications.

corona

Whether to update spec by outliers according to corona_outliers. FALSE or NULL means no update. TRUE or "ssb" means update.

...

Further x13 parameters (currently only parameter userdefined is additional parameter to x13).

pickmdl_method

crit_selection parameter or one of the two extra possibilities, "first_automdl" or "first_tryautomdl". In both cases the crit_selection parameter is "first" and the automdl model is added as the last pickmdl model.

"first_automdl": The automdl model is chosen whenever no pickmdl model is ok. In other words, the star parameter changes.

"first_tryautomdl": When no pickmdl model is ok: The automdl model is chosen if this model is ok, otherwise the star model is chosen.

star

crit_selection parameter

when_star

crit_selection parameter

when_automdl

Function to be called when automdl since no pickmdl model ok. Supply NULL to do nothing.

when_finalnotok

Function to be called, e.g. warning, when final run with final model is not ok. Supply NULL to do nothing. See crit_ok.

identification_end

To shorten the series before runs used to identify (arima) parameters. That is, the series is shortened by window(series, end = identification_end).

identification_estimate.to

To set x13_spec parameter estimate.to before runs used to identify (arima) parameters. This is an alternative to identification_end.

identify_t_filter

When TRUE, Henderson trend filter is identified by the shortened (see above) series.

identify_s_filter

When TRUE, Seasonal moving average filter is identified by the shortened series.

identify_outliers

When TRUE, Outliers are identified by the shortened series.

identify_arima_mu

When TRUE, arima.mu is identified by the shortened series (see arima_mu).

automdl.enabled

When TRUE, automdl is performed instead of pickmdl. If spec is a list of several objects as outputted from x13_spec_pickmdl, only first object is used.

fastfirst

When TRUE and when pickmdl with crit_selection parameter "first", only as many models as needed are run. This affects the output when output = "all".

verbose

Printing information to console when TRUE.

output

One of "sa" (default), "spec" (final spec), "sa_spec" (both) and "all". See examples.

add_comment

When TRUE, a comment attribute (character vector with ok, ok_final and mdl_nr) will be added to the x13 output object. Use comment to get the attribute or ok to get the attribute converted to a list.

Value

By default an x13 output object, or otherwise a list as specified by parameter output.

Examples

myseries <- pickmdl_data("myseries")

spec_a  <- x13_spec(spec = "RSA3", transform.function = "Log")

a <- x13_pickmdl(myseries, spec_a, verbose = TRUE)
#>  p  d  q bp bd bq 
#>  0  1  2  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.
comment(a)
#>       ok ok_final   mdl_nr 
#>   "TRUE"   "TRUE"      "2" 
ok(a)
#> $ok
#> [1] TRUE
#> 
#> $ok_final
#> [1] TRUE
#> 
#> $mdl_nr
#> [1] 2
#> 
unlist(ok(a))
#>       ok ok_final   mdl_nr 
#>        1        1        2 
a$regarima
#> y = regression model + arima (0, 1, 2, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -1.1853      0.065
#> Theta(2)    0.4694      0.066
#> BTheta(1)  -0.8489      0.050
#> 
#> 
#> Residual standard error: 0.05928 on 184 degrees of freedom
#> Log likelihood =   256, aic =  1207 aicc =  1207, bic(corrected for length) = -5.568
#> 

a2 <- x13_pickmdl(myseries, spec_a, identification_end = c(2014, 2))
a2$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.8366      0.041
#> BTheta(1)  -0.9039      0.045
#> 
#> 
#> Residual standard error: 0.06579 on 185 degrees of freedom
#> Log likelihood = 234.2, aic =  1249 aicc =  1249, bic(corrected for length) = -5.387
#> 

# As above, another way
a3 <- x13_pickmdl(myseries, spec_a, identification_estimate.to = "2014-03-01")
a3$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.8366      0.041
#> BTheta(1)  -0.9039      0.045
#> 
#> 
#> Residual standard error: 0.06579 on 185 degrees of freedom
#> Log likelihood = 234.2, aic =  1249 aicc =  1249, bic(corrected for length) = -5.387
#> 

a4 <- x13_automdl(myseries, spec_a, identification_end = c(2014, 2))
a4$regarima
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.8934      0.083
#> Phi(2)      0.5646      0.074
#> Theta(1)   -0.3434      0.094
#> BTheta(1)  -0.8074      0.053
#> 
#> 
#> Residual standard error: 0.0559 on 183 degrees of freedom
#> Log likelihood = 268.2, aic =  1185 aicc =  1185, bic(corrected for length) = -5.657
#> 

# As above, another way
spec_a_single  <- x13_spec(spec = "RSA3", transform.function = "Log")
a5 <- x13_automdl(myseries, spec_a_single, identification_estimate.to = "2014-03-01")
a5$regarima
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.8934      0.083
#> Phi(2)      0.5646      0.074
#> Theta(1)   -0.3434      0.094
#> BTheta(1)  -0.8074      0.053
#> 
#> 
#> Residual standard error: 0.0559 on 183 degrees of freedom
#> Log likelihood = 268.2, aic =  1185 aicc =  1185, bic(corrected for length) = -5.657
#> 

allvar <- pickmdl_data("allvar")

spec_b <- x13_spec(
            spec = "RSA3", transform.function = "Log",
            usrdef.varEnabled = TRUE, 
            usrdef.varType = c("Calendar", "Calendar"), 
            usrdef.var = allvar, 
            outlier.enabled = FALSE, 
            usrdef.outliersEnabled = TRUE,
            usrdef.outliersType = rep("LS", 20), 
            usrdef.outliersDate = c("2009-01-01", "2016-01-01", 
                                    "2020-03-01", "2020-04-01", "2020-05-01", 
                                    "2020-06-01", "2020-07-01", "2020-08-01", 
                                    "2020-09-01", "2020-10-01", "2020-11-01", 
                                    "2020-12-01", "2021-01-01", "2021-02-01",
                                    "2021-03-01", "2021-04-01", "2021-05-01",
                                    "2021-06-01", "2021-07-01", "2021-08-01"))
b <- x13_pickmdl(myseries, spec_b, identification_end = c(2020, 2))                                     
b$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5815      0.064
#> BTheta(1)  -0.3136      0.073
#> 
#>                Estimate Std. Error
#> arb_dag       0.0129544      0.000
#> skuddar       0.0380289      0.009
#> LS (1-2009)  -0.0810765      0.015
#> LS (1-2016)   0.0080841      0.015
#> LS (3-2020)  -0.0040775      0.023
#> LS (4-2020)  -0.0325380      0.027
#> LS (5-2020)   0.0244522      0.027
#> LS (6-2020)  -0.0212278      0.027
#> LS (7-2020)   0.0057234      0.027
#> LS (8-2020)  -0.0133100      0.026
#> LS (9-2020)   0.0035141      0.022
#> LS (10-2020)  0.0119548      0.026
#> LS (11-2020)  0.0007714      0.027
#> LS (12-2020)  0.0087747      0.027
#> LS (1-2021)   0.0138112      0.027
#> LS (2-2021)  -0.0012638      0.028
#> LS (3-2021)  -0.0119383      0.031
#> LS (4-2021)  -0.0256174      0.032
#> LS (5-2021)  -0.0052522      0.032
#> LS (6-2021)   0.0064868      0.033
#> LS (7-2021)  -0.0533469      0.032
#> LS (8-2021)   0.0482770      0.029
#> 
#> 
#> Residual standard error: 0.02153 on 163 degrees of freedom
#> Log likelihood =   454, aic = 853.3 aicc = 861.3, bic(corrected for length) = -7.008
#> 

# automdl instead  
b1 <- x13_automdl(myseries, spec_b, identification_end = c(2020, 2))
b1$regarima
#> y = regression model + arima (0, 1, 1, 1, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> BPhi(1)    -0.4571      0.114
#> Theta(1)   -0.5902      0.063
#> BTheta(1)  -0.7409      0.084
#> 
#>               Estimate Std. Error
#> arb_dag       0.012763      0.000
#> skuddar       0.036786      0.009
#> LS (1-2009)  -0.077338      0.015
#> LS (1-2016)   0.010572      0.015
#> LS (3-2020)  -0.002627      0.023
#> LS (4-2020)  -0.042349      0.026
#> LS (5-2020)   0.028011      0.026
#> LS (6-2020)  -0.017489      0.026
#> LS (7-2020)   0.008144      0.026
#> LS (8-2020)  -0.017247      0.025
#> LS (9-2020)   0.005831      0.021
#> LS (10-2020)  0.012817      0.025
#> LS (11-2020)  0.002899      0.026
#> LS (12-2020)  0.011755      0.026
#> LS (1-2021)   0.004380      0.026
#> LS (2-2021)   0.006393      0.027
#> LS (3-2021)  -0.009891      0.031
#> LS (4-2021)  -0.039036      0.032
#> LS (5-2021)  -0.002961      0.032
#> LS (6-2021)   0.011332      0.032
#> LS (7-2021)  -0.045436      0.032
#> LS (8-2021)   0.039530      0.029
#> 
#> 
#> Residual standard error: 0.02089 on 162 degrees of freedom
#> Log likelihood = 459.1, aic = 845.1 aicc = 853.8, bic(corrected for length) = -7.041
#> 

# effect of identify_t_filter and identify_s_filter
set.seed(1)
rndseries <- ts(rep(1:12, 20) + (1 + (1:240)/20) * runif(240) + 0.5 * c(rep(1, 120), (1:120)^2), 
                frequency = 12, start = c(2000, 1))
spec_c <- x13_spec(outlier.enabled = FALSE)               
c1 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12))    
c1$decomposition
#> Monitoring and Quality Assessment Statistics:
#>       M stats
#> M(1)    0.200
#> M(2)    0.024
#> M(3)    0.020
#> M(4)    0.139
#> M(5)    0.268
#> M(6)    0.783
#> M(7)    1.307
#> M(8)    0.915
#> M(9)    0.730
#> M(10)   0.120
#> M(11)   0.120
#> Q       0.472
#> Q-M2    0.535
#> 
#> Final filters: 
#> Seasonal filter:  3x3
#> Trend filter:  13 terms Henderson moving average
c2 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE) 
c2$decomposition
#> Monitoring and Quality Assessment Statistics:
#>       M stats
#> M(1)    0.248
#> M(2)    0.023
#> M(3)    0.000
#> M(4)    0.139
#> M(5)    0.289
#> M(6)    0.768
#> M(7)    1.288
#> M(8)    0.933
#> M(9)    0.731
#> M(10)   0.060
#> M(11)   0.059
#> Q       0.470
#> Q-M2    0.533
#> 
#> Final filters: 
#> Seasonal filter:  3x3
#> Trend filter:  23-Henderson
c3 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE, 
                  identify_s_filter = TRUE)     
c3$decomposition                       
#> Monitoring and Quality Assessment Statistics:
#>       M stats
#> M(1)    0.378
#> M(2)    0.032
#> M(3)    0.069
#> M(4)    0.338
#> M(5)    0.317
#> M(6)    0.751
#> M(7)    1.270
#> M(8)    0.830
#> M(9)    0.735
#> M(10)   0.093
#> M(11)   0.092
#> Q       0.531
#> Q-M2    0.592
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  23-Henderson

# Warning when transform.function = "None"
spec_d  <- x13_spec(spec = "RSA3", transform.function = "None")
d <- x13_pickmdl(myseries, spec_d, verbose = TRUE)
#> Warning: No model is ok according to criteria
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.

# Warning avoided (when_star) and 2nd (star) model selected 
d2 <- x13_pickmdl(myseries, spec_d, star = 2, when_star = NULL, verbose = TRUE)
#>  p  d  q bp bd bq 
#>  0  1  2  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.

# automdl since no pickmdl model ok, but still not ok 
d3 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_automdl", verbose = TRUE)
#> Warning: No model is ok according to criteria
#>  p  d  q bp bd bq 
#>  3  0  1  0  1  1 
#> automdl since no pickmdl model ok
#> outlier.from updated: 2021-10-01  No new outliers.

# airline model (star) since automdl also not ok 
d4 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_tryautomdl", verbose = TRUE,
                  when_finalnotok = warning) # also finalnotok warning
#> Warning: No model is ok according to criteria
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.
#> Warning: FINAL RUN NOT OK

# As a2, with output = "all"
k <- x13_pickmdl(myseries, spec_b, identification_end = c(2010, 2), output = "all",
                 fastfirst = FALSE) # With TRUE only one model in this case 
k$sa$decomposition  # As a2$decomposition 
#> Monitoring and Quality Assessment Statistics:
#>       M stats
#> M(1)    0.017
#> M(2)    0.016
#> M(3)    0.411
#> M(4)    0.174
#> M(5)    0.512
#> M(6)    0.553
#> M(7)    0.067
#> M(8)    0.176
#> M(9)    0.156
#> M(10)   0.246
#> M(11)   0.243
#> Q       0.225
#> Q-M2    0.251
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
k$mdl_nr            # index of selected model used to identify parameters
#> [1] 1
k$sa_mult[[k$mdl_nr]]$decomposition  # decomposition for model to identify
#> Monitoring and Quality Assessment Statistics:
#>       M stats
#> M(1)    0.016
#> M(2)    0.016
#> M(3)    0.611
#> M(4)    0.000
#> M(5)    0.616
#> M(6)    0.047
#> M(7)    0.092
#> M(8)    0.000
#> M(9)    0.000
#> M(10)   0.000
#> M(11)   0.000
#> Q       0.168
#> Q-M2    0.194
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
k$crit_tab          # Table of criteria 
#>           crit1     crit2      crit3    m_aic
#> [1,] 0.02036440 1.0000000 -0.5730277 243.7063
#> [2,] 0.02045199 1.0000000 -0.5784995 245.6822
#> [3,] 0.02054858 1.0000000  0.0000000 246.7835
#> [4,] 0.01979950 0.9533652 -0.9623400 244.9268
#> [5,] 0.01809707 0.9860155 -0.6897453 243.7226

# Effect of identify_outliers (TRUE is default)
m1 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = FALSE)
m2 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, output = "all")
#>  p  d  q bp bd bq 
#>  2  1  0  0  1  1 
#> outlier.from updated: 2010-03-01  New outliers: 2005-04, 2008-04 
m3 <- x13_pickmdl(myseries, m2$spec, identification_end = c(2018, 2), identify_outliers = TRUE, 
                  verbose = TRUE)
#>  p  d  q bp bd bq 
#>  2  1  0  0  1  1 
#> outlier.from updated: 2018-03-01  New outliers: 2011-06, 2017-04 

m1$regarima
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      1.0894      0.052
#> Phi(2)      0.7085      0.052
#> BTheta(1)  -0.7767      0.055
#> 
#>             Estimate Std. Error
#> TC (6-2011)  -0.1308      0.034
#> AO (4-2005)   0.1365      0.043
#> 
#> 
#> Residual standard error: 0.05412 on 182 degrees of freedom
#> Log likelihood = 274.9, aic =  1173 aicc =  1174, bic(corrected for length) = -5.694
#> 
m2$sa$regarima
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      1.0899      0.052
#> Phi(2)      0.7104      0.052
#> BTheta(1)  -0.7707      0.055
#> 
#>             Estimate Std. Error
#> AO (4-2005)   0.1480      0.042
#> TC (4-2008)   0.0963      0.034
#> TC (6-2011)  -0.1314      0.033
#> 
#> 
#> Residual standard error: 0.05298 on 181 degrees of freedom
#> Log likelihood = 279.1, aic =  1167 aicc =  1168, bic(corrected for length) = -5.709
#> 
m3$regarima
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      1.0949      0.052
#> Phi(2)      0.7163      0.051
#> BTheta(1)  -0.7557      0.056
#> 
#>             Estimate Std. Error
#> AO (4-2005)  0.14736      0.041
#> TC (4-2008)  0.09449      0.033
#> TC (6-2011) -0.13113      0.032
#> TC (4-2017) -0.09238      0.033
#> 
#> 
#> Residual standard error: 0.05193 on 180 degrees of freedom
#> Log likelihood = 283.1, aic =  1161 aicc =  1162, bic(corrected for length) = -5.721
#> 

# With corona outliers (even possible when series is not long enough) 
m4 <- x13_pickmdl(myseries, spec_a, verbose = TRUE, corona = TRUE)
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.
m4$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.8357      0.043
#> BTheta(1)  -0.9999      0.053
#> 
#>               Estimate Std. Error
#> LS (3-2020)   0.053680      0.066
#> LS (4-2020)  -0.159863      0.086
#> LS (5-2020)   0.041034      0.086
#> LS (6-2020)   0.013213      0.086
#> LS (7-2020)   0.105120      0.086
#> LS (8-2020)  -0.182598      0.086
#> LS (9-2020)   0.110802      0.086
#> LS (10-2020) -0.019992      0.086
#> LS (11-2020) -0.002362      0.086
#> LS (12-2020)  0.092031      0.086
#> LS (1-2021)  -0.180451      0.086
#> LS (2-2021)   0.088556      0.086
#> LS (3-2021)   0.069173      0.086
#> LS (4-2021)  -0.107602      0.086
#> LS (5-2021)  -0.079352      0.086
#> LS (6-2021)   0.131609      0.086
#> LS (7-2021)  -0.044631      0.086
#> LS (8-2021)  -0.033696      0.086
#> LS (9-2021)   0.069691      0.086
#> 
#> 
#> Residual standard error: 0.0601 on 166 degrees of freedom
#> Log likelihood = 244.3, aic =  1267 aicc =  1273, bic(corrected for length) = -5.039
#> 
m5 <- x13_pickmdl(myseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, corona = TRUE) 
#>  p  d  q bp bd bq 
#>  2  1  0  0  1  1 
#> outlier.from updated: 2010-03-01  New outliers: 2005-04, 2008-04 
m5$regarima 
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      1.0897      0.052
#> Phi(2)      0.7500      0.051
#> BTheta(1)  -0.8231      0.048
#> 
#>               Estimate Std. Error
#> LS (3-2020)   0.078956      0.050
#> LS (4-2020)  -0.150119      0.074
#> LS (5-2020)   0.027823      0.077
#> LS (6-2020)   0.001453      0.078
#> LS (7-2020)   0.084143      0.086
#> LS (8-2020)  -0.157041      0.089
#> LS (9-2020)   0.109967      0.089
#> LS (10-2020) -0.030510      0.091
#> LS (11-2020) -0.002315      0.093
#> LS (12-2020)  0.099073      0.094
#> LS (1-2021)  -0.164471      0.094
#> LS (2-2021)   0.074665      0.095
#> LS (3-2021)   0.089341      0.095
#> LS (4-2021)  -0.106637      0.095
#> LS (5-2021)  -0.097606      0.096
#> LS (6-2021)   0.131929      0.096
#> LS (7-2021)  -0.074990      0.096
#> LS (8-2021)  -0.006976      0.096
#> LS (9-2021)   0.074627      0.096
#> AO (4-2005)   0.150159      0.038
#> TC (4-2008)   0.093193      0.031
#> TC (6-2011)  -0.135924      0.031
#> TC (4-2017)  -0.097332      0.032
#> AO (3-2019)   0.111171      0.038
#> AO (4-2016)   0.104861      0.037
#> 
#> 
#> Residual standard error: 0.04581 on 159 degrees of freedom
#> Log likelihood = 304.9, aic =  1159 aicc =  1170, bic(corrected for length) = -5.387
#> 

 
###########  quarterly series  #############
   
qseries <- pickmdl_data("qseries")    

# Effect of identify_outliers (TRUE is default)
q1 <- x13_pickmdl(qseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = FALSE)
q2 <- x13_pickmdl(qseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, output = "all")
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2010-07-01  New outliers: 2008-07 
q3 <- x13_pickmdl(qseries, q2$spec, identification_end = c(2018, 2), identify_outliers = TRUE, 
                  verbose = TRUE)
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2018-07-01  New outliers: 2014-04 

q1$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.2941      0.125
#> BTheta(1)  -0.7877      0.104
#> 
#> 
#> Residual standard error: 0.04514 on 59 degrees of freedom
#> Log likelihood = 102.1, aic = 369.5 aicc = 369.9, bic(corrected for length) = -6.063
#> 
q2$sa$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.4205      0.120
#> BTheta(1)  -0.7619      0.106
#> 
#>               Estimate Std. Error
#> LS (III-2008)   -0.134      0.037
#> 
#> 
#> Residual standard error: 0.04142 on 58 degrees of freedom
#> Log likelihood = 107.6, aic = 360.6 aicc = 361.3, bic(corrected for length) = -6.168
#> 
q3$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.3971      0.122
#> BTheta(1)  -0.7461      0.106
#> 
#>               Estimate Std. Error
#> LS (III-2008) -0.13357      0.035
#> AO (II-2014)   0.09058      0.031
#> 
#> 
#> Residual standard error: 0.03882 on 57 degrees of freedom
#> Log likelihood = 111.7, aic = 354.3 aicc = 355.4, bic(corrected for length) = -6.231
#> 

# With corona outliers (even possible when series is not long enough) 
q4 <- x13_pickmdl(qseries, spec_a, verbose = TRUE, corona = TRUE)
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2021-10-01  No new outliers.
q4$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.2713      0.135
#> BTheta(1)  -0.9997      0.087
#> 
#>                Estimate Std. Error
#> LS (I-2020)   -0.031164      0.043
#> LS (II-2020)  -0.038923      0.044
#> LS (III-2020) -0.064265      0.044
#> LS (IV-2020)   0.088448      0.044
#> LS (I-2021)   -0.002755      0.045
#> LS (II-2021)  -0.116141      0.044
#> LS (III-2021)  0.053282      0.044
#> 
#> 
#> Residual standard error: 0.03839 on 52 degrees of freedom
#> Log likelihood = 108.5, aic = 370.7 aicc = 375.1, bic(corrected for length) = -5.921
#> 
q5 <- x13_pickmdl(qseries, x13_spec("RSA3", outlier.usedefcv = FALSE, outlier.cv = 3), 
                  identification_end = c(2010, 2), identify_outliers = TRUE, 
                  verbose = TRUE, corona = TRUE) 
#>  p  d  q bp bd bq 
#>  0  1  1  0  1  1 
#> outlier.from updated: 2010-07-01  New outliers: 2008-07 
q5$regarima 
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.3110      0.136
#> BTheta(1)  -0.8427      0.084
#> 
#>                Estimate Std. Error
#> LS (I-2020)   -0.030835      0.035
#> LS (II-2020)  -0.021382      0.036
#> LS (III-2020) -0.082770      0.036
#> LS (IV-2020)   0.077653      0.036
#> LS (I-2021)    0.002271      0.037
#> LS (II-2021)  -0.098601      0.037
#> LS (III-2021)  0.034777      0.037
#> LS (III-2008) -0.133177      0.032
#> AO (II-2014)   0.093274      0.027
#> AO (II-2011)   0.085193      0.027
#> 
#> 
#> Residual standard error: 0.03105 on 49 degrees of freedom
#> Log likelihood = 124.8, aic = 344.2 aicc = 351.8, bic(corrected for length) = -6.146
#> 

# Demonstrate strange behavior of x13 with TC at end. Updating outlier.from matters.
# Explains why TC (III-2021) is in q4 and not in q5 
q11 <- x13(window(qseries, end = c(2020, 1)), 
    spec = x13_spec(spec = "RSA3", transform.function = "Log", 
    usrdef.outliersEnabled = TRUE, usrdef.outliersType = "TC", 
    usrdef.outliersDate = "2020-01-01"))  
q12 <- x13(window(qseries, end = c(2020, 1)), # same with outlier.from 
    spec = x13_spec(spec = "RSA3", transform.function = "Log", 
    usrdef.outliersEnabled = TRUE, usrdef.outliersType = "TC", 
    usrdef.outliersDate = "2020-01-01", outlier.from = "2020-04-01")) 
q11$regarima 
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.2676      0.135
#> BTheta(1)  -0.9100      0.093
#> 
#> 
#> Residual standard error: 0.04206 on 53 degrees of freedom
#> Log likelihood = 94.55, aic = 327.9 aicc = 328.4, bic(corrected for length) = -6.194
#> 
q12$regarima
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.2693      0.136
#> BTheta(1)  -0.9248      0.092
#> 
#>             Estimate Std. Error
#> TC (I-2020) -0.03069      0.043
#> 
#> 
#> Residual standard error: 0.04166 on 52 degrees of freedom
#> Log likelihood = 94.81, aic = 329.4 aicc = 330.2, bic(corrected for length) = -6.141
#>