Multiple X-13ARIMA model specifications

x13_spec is run multiple times with input for multiple arima models.

Usage

x13_spec_pickmdl(
  ...,
  arima.p = c(0, 0, 2, 0, 2),
  arima.d = c(1, 1, 1, 2, 1),
  arima.q = c(1, 2, 0, 2, 2),
  arima.bp = 0,
  arima.bd = 1,
  arima.bq = 1,
  automdl.enabled = FALSE
)

Arguments

...

x13_spec parameters

arima.p

x13_spec parameter as vector

arima.d

x13_spec parameter as vector

arima.q

x13_spec parameter as vector

arima.bp

x13_spec parameter, possibly as vector

arima.bd

x13_spec parameter, possibly as vector

arima.bq

x13_spec parameter, possibly as vector

automdl.enabled

x13_spec parameter

Value

List of several x13_spec output objects

Details

This function behaves like x13_spec except that some of the parameters may be vectors. These vectors must be the same length.

Examples

spec5 <- x13_spec_pickmdl(spec = "RSA3", transform.function = "Log")